requirements.txt
numpy>=1.23
pandas>=1.5
matplotlib>=3.7
seaborn>=0.12

scikit-learn>=1.2

Pillow>=9.5
opencv-python>=4.8

torch>=2.0
torchvision>=0.15
torchaudio>=2.0

tqdm>=4.65

jupyter>=1.0
notebook>=7.0
